reading-companion-backend/pyproject.toml
[project]
name = "reading-companion"
version = "0.1.0"
description = "阅读伴侣 Agent - 一个读得更多的共读者"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "tavily-python>=0.3.0",
    "pydantic>=2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "socksio>=1.0.0",
    "ebooklib>=0.5",
    "pymupdf>=1.23",
    "mobi>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
